+++ buildSrc/src/main/kotlin/Publication.kt
@@ -1,7 +1,7 @@
 object Publication {
     const val GROUP = "net.redwarp.gif"
-    const val VERSION_CODE = 3
-    const val VERSION_NAME = "0.2.0"
+    const val VERSION_CODE = 4
+    const val VERSION_NAME = "0.2.1"
 
     object Pom {
         const val URL = "https://github.com/redwarp/gifdecoder"